projects/accounts-api/requirements.txt
asyncpg==0.30.0
black==24.10.0
email-validator==2.2.0
fastapi==0.115.7
passlib==1.7.4
pip-chill==1.0.3
tortoise-orm==0.24.0
uvicorn==0.34.0
